requirements/dotenv.txt

# https://github.com/theskumar/python-dotenv
python-dotenv==0.14.0;python_version>='3.6.1'
